chore(core): dedicated metadata module
Move operations related to metadata fields which aren't used for visual encoding (text descriptions, urls for links, etc) into a dedicated module instead of intertwining them with the data handling. This should make it easier to improve.
