+++ source/data.js
@@ -7,8 +7,9 @@ import {
   encodingField,
   encodingValue,
 } from './encodings.js';
+import { metadata } from './metadata.js';
 import { feature } from './feature.js';
-import { identity, isDiscrete, missingSeries, nested, values } from './helpers.js';
+import { identity, missingSeries, nested, values } from './helpers.js';
 import { memoize } from './memoize.js';
 import { parseTime } from './time.js';
 
@@ -124,86 +125,6 @@ const sort = (data) => {
 };
 
 /**
- * copy URL from raw data points to aggregate data point
- * @param {array} aggregated nested aggregate data points
- * @param {array} raw raw data points
- * @param {object} s Vega Lite specification
- * @returns {object} nested aggregate data points with URLs added
- */
-const transplantStackedBarMetadata = (aggregated, raw, s) => {
-  const createMatcher = (key) => {
-    const matcher = (aggregatedItem, raw) => {
-      const laneChannel = encodingChannelCovariateCartesian(s);
-      const keys = {
-        lane: encodingField(s, laneChannel),
-        series: encodingField(s, 'color'),
-      };
-      const matches = raw
-        .filter((rawItem) => {
-          let seriesMatch;
-
-          // single-color categorical charts are still plotted using separate series nodes
-          if (feature(s).hasColor() || isDiscrete(s, encodingChannelCovariate(s))) {
-            seriesMatch = aggregatedItem[keys.series] === rawItem[keys.series];
-          } else {
-            seriesMatch = true;
-          }
-
-          const laneMatch =
-            aggregatedItem[keys.lane] &&
-            rawItem[keys.lane] &&
-            aggregatedItem[keys.lane]?.toString() === rawItem[keys.lane]?.toString();
-          const hasField = !!rawItem[key];
-
-          return seriesMatch && laneMatch && hasField;
-        })
-        .map((item) => item[key])
-        .filter(Boolean);
-
-      return matches;
-    };
-
-    return matcher;
-  };
-
-  const results = [...aggregated];
-
-  results.forEach((series, i) => {
-    series.forEach((item, j) => {
-      const lookup = {};
-
-      if (encodingField(s, 'color')) {
-        lookup[encodingField(s, 'color')] = series.key;
-      }
-
-      if (encodingField(s, encodingChannelCovariate(s))) {
-        lookup[encodingField(s, encodingChannelCovariate(s))] = item.data.key;
-      }
-
-      const channels = ['href', 'description', 'tooltip'];
-      const fields = channels.map((channel) => encodingField(s, channel));
-
-      fields.forEach((field) => {
-        const value = transplantFields([lookup], s.data.values, createMatcher(field), field).pop()[
-          field
-        ];
-
-        if (value) {
-          // the reference to the datum object is shared across marks by the d3 layout generator, so
-          // this needs to be set as a non-enumerable property on the array, which is a bit odd but
-          // also how the datum object is already stored
-          Object.defineProperty(results[i][j], field, {
-            value,
-          });
-        }
-      });
-    });
-  });
-
-  return results;
-};
-
-/**
  * retrieve a numerical value for the stack layout
  * @param {object} d datum
  * @param {string} key property
@@ -244,7 +165,8 @@ const _stackData = (s) => {
   }
   const sorted = sort(stacked);
 
-  return transplantStackedBarMetadata(sorted, values(s), s);
+  return metadata(sorted, values(s), s);
+
 };
 
 /**
@@ -268,35 +190,8 @@ const _circularData = (s) => {
 
   results = summed;
 
-  const createMatcher = (channel) => {
-    return (aggregatedItem, raw) => {
-      return raw
-        .filter((rawItem) => {
-          const currentAggregateDatumField = aggregatedItem.key.toString();
-          const rawDatumField = rawItem[encodingField(s, 'color')].toString();
+  return metadata(results, values(s), s);
 
-          return currentAggregateDatumField === rawDatumField;
-        })
-        .map((item) => {
-          return item[encodingField(s, channel)];
-        });
-    };
-  };
-
-  const channels = ['href', 'description', 'tooltip'];
-
-  channels.forEach((channel) => {
-    if (s.encoding[channel]) {
-      results = transplantFields(
-        [...results],
-        s.data.values,
-        createMatcher(channel),
-        encodingField(s, channel),
-      );
-    }
-  });
-
-  return results;
 };
 
 /**
@@ -363,41 +258,6 @@ const textData = (s) => {
   return s.data.values;
 };
 
-/**
- * move properties from an array of source
- * values to an aggregate
- * @param {array} aggregated aggregated data points
- * @param {array} raw individual data points
- * @param {function} matcher find matching values
- * @param {string} key property name for transplanted field
- * @returns {array} aggregated data points with transplanted field attached
- */
-const transplantFields = (aggregated, raw, matcher, key) => {
-  if (!Array.isArray(aggregated)) {
-    throw new Error(`cannot transplant ${key} field, aggregated data is not an array`);
-  }
-
-  if (!Array.isArray(raw)) {
-    throw new Error(`cannot transplant ${key} field, raw data is not an array`);
-  }
-
-  if (typeof matcher !== 'function') {
-    throw new Error(`cannot transplant ${key} field, matcher is not a function`);
-  }
-
-  return aggregated.map((item) => {
-    const matches = matcher(item, raw);
-    const result = { ...item };
-    const allMatch = matches.every((item) => item === matches[0]);
-
-    if (matches.length > 0 && allMatch) {
-      result[key] = matches[0];
-    }
-
-    return result;
-  });
-};
-
 const pointData = identity;
 
 /**
@@ -419,4 +279,4 @@ const data = (s) => {
   }
 };
 
-export { data, pointData, sumByCovariates, transplantFields };
+export { data, pointData, sumByCovariates };
